+++ src/subscription_manager/gui/mysubstab.py
@@ -1,3 +1,5 @@
+from __future__ import print_function, division, absolute_import
+
 #
 # Copyright (c) 2010 Red Hat, Inc.
 #
@@ -12,7 +14,6 @@
 # granted to use or replicate Red Hat trademarks that are incorporated
 # in this software or its documentation.
 #
-
 import gettext
 import os
 from datetime import datetime
@@ -141,7 +142,7 @@ class MySubscriptionsTab(widgets.SubscriptionManagerTab):
         if not response:
             return
 
-        serial = long(selection['serial'])
+        serial = int(selection['serial'])
 
         if self.identity.is_valid():
             self.pb = progress.Progress(_("Removing"),
@@ -269,7 +270,7 @@ class MySubscriptionsTab(widgets.SubscriptionManagerTab):
         if not serial:
             return
 
-        cert = self.entitlement_dir.find(long(serial))
+        cert = self.entitlement_dir.find(int(serial))
         if not cert:
             return
 
@@ -380,7 +381,7 @@ class MySubscriptionsTab(widgets.SubscriptionManagerTab):
         if (len(full_set) == 0):
             return 100
         else:
-            return (float(len(subset)) / len(full_set)) * 100
+            return (len(subset) / len(full_set)) * 100
 
     def _get_installed(self, products):
         installed_dir = self.product_dir
